Revert "dpa_offload: Reduce the size of the IPv4 header options field"
This reverts commit 6b31fa5a03957a3bee4414a06ac02c1a8233fb2c.

+++ include/linux/fsl_dpa_offload.h
@@ -162,7 +162,7 @@ struct ipv4_header {
 	uint8_t				*options;
 
 	/* Size of IPv4 options buffer. Zero for no options. */
-	uint8_t				options_size;
+	unsigned int			options_size;
 };
 
 /* Description of the VLAN header */
